move ShaderBuilder and style parsing to ol/render/webgl
Renamed styleparser.js to style.js to match the convention in ol/canvas/render

Split up utils file by topic:
* compileUtil.js is used to compile expressions
* encodeUtil.js is for encoding ids as colors
* bufferUtil.js is used for filling buffers

Also removed unused function getBlankImageData
