+++ config/tsconfig-strict.json
@@ -185,7 +185,11 @@
     "../src/ol/render/webgl/constants.js",
     "../src/ol/render/webgl/MixedGeometryBatch.js",
     "../src/ol/render/webgl/renderinstructions.js",
-    "../src/ol/render/webgl/utils.js",
+    "../src/ol/render/webgl/bufferUtil.js",
+    "../src/ol/render/webgl/compileUtil.js",
+    "../src/ol/render/webgl/encodeUtil.js",
+    "../src/ol/render/webgl/ShaderBuilder.js",
+    "../src/ol/render/webgl/style.js",
     "../src/ol/render/webgl/VectorStyleRenderer.js",
     "../src/ol/renderer/canvas/ImageLayer.js",
     "../src/ol/renderer/canvas/Layer.js",
@@ -290,8 +294,6 @@
     "../src/ol/webgl/PaletteTexture.js",
     "../src/ol/webgl/PostProcessingPass.js",
     "../src/ol/webgl/RenderTarget.js",
-    "../src/ol/webgl/ShaderBuilder.js",
-    "../src/ol/webgl/style.js",
     "../src/ol/webgl/TileGeometry.js",
     "../src/ol/webgl/TileTexture.js",
     "../src/ol/worker/webgl.js",
