+++ src/ol/render/webgl/ShaderBuilder.js
@@ -1,10 +1,10 @@
 /**
  * Class for generating shaders from literal style objects
- * @module ol/webgl/ShaderBuilder
+ * @module ol/render/webgl/ShaderBuilder
  */
-import {colorToGlsl, numberToGlsl, stringToGlsl} from '../expr/gpu.js';
-import {LINESTRING_ANGLE_COSINE_CUTOFF} from '../render/webgl/utils.js';
-import {createDefaultStyle} from '../style/flat.js';
+import {colorToGlsl, numberToGlsl, stringToGlsl} from '../../expr/gpu.js';
+import {createDefaultStyle} from '../../style/flat.js';
+import {LINESTRING_ANGLE_COSINE_CUTOFF} from './bufferUtil.js';
 
 export const COMMON_HEADER = `#ifdef GL_FRAGMENT_PRECISION_HIGH
 precision highp float;
