+++ src/ol/render/webgl/bufferUtil.js
@@ -1,5 +1,6 @@
 /**
- * @module ol/render/webgl/utils
+ * Utilities for filling WebGL buffers
+ * @module ol/render/webgl/bufferUtil
  */
 import earcut from 'earcut';
 import {clamp} from '../../math.js';
@@ -379,126 +380,4 @@ export function writePolygonTrianglesToBuffers(
   }
 
   return instructionsIndex + verticesCount * instructionsPerVertex;
-}
-
-/**
- * Returns a texture of 1x1 pixel, white
- * @private
- * @return {ImageData} Image data.
- */
-export function getBlankImageData() {
-  const canvas = document.createElement('canvas');
-  const image = canvas.getContext('2d').createImageData(1, 1);
-  image.data[0] = 255;
-  image.data[1] = 255;
-  image.data[2] = 255;
-  image.data[3] = 255;
-  return image;
-}
-
-/**
- * Generates a color array based on a numerical id
- * Note: the range for each component is 0 to 1 with 256 steps
- * @param {number} id Id
- * @param {Array<number>} [array] Reusable array
- * @return {Array<number>} Color array containing the encoded id
- */
-export function colorEncodeId(id, array) {
-  array = array || [];
-  const radix = 256;
-  const divide = radix - 1;
-  array[0] = Math.floor(id / radix / radix / radix) / divide;
-  array[1] = (Math.floor(id / radix / radix) % radix) / divide;
-  array[2] = (Math.floor(id / radix) % radix) / divide;
-  array[3] = (id % radix) / divide;
-  return array;
-}
-
-/**
- * Reads an id from a color-encoded array
- * Note: the expected range for each component is 0 to 1 with 256 steps.
- * @param {Array<number>} color Color array containing the encoded id
- * @return {number} Decoded id
- */
-export function colorDecodeId(color) {
-  let id = 0;
-  const radix = 256;
-  const mult = radix - 1;
-  id += Math.round(color[0] * radix * radix * radix * mult);
-  id += Math.round(color[1] * radix * radix * mult);
-  id += Math.round(color[2] * radix * mult);
-  id += Math.round(color[3] * mult);
-  return id;
-}
-
-/**
- * @typedef {import('./VectorStyleRenderer.js').AsShaders} StyleAsShaders
- */
-/**
- * @typedef {import('./VectorStyleRenderer.js').AsRule} StyleAsRule
- */
-
-/**
- * Takes in either a Flat Style or an array of shaders (used as input for the webgl vector layer classes)
- * and breaks it down into separate styles to be used by the VectorStyleRenderer class.
- * @param {import('../../style/flat.js').FlatStyleLike | Array<StyleAsShaders> | StyleAsShaders} style Flat style or shaders
- * @return {Array<StyleAsShaders | StyleAsRule>} Separate styles as shaders or rules with a single flat style and a filter
- */
-export function breakDownFlatStyle(style) {
-  // possible cases:
-  // - single shader
-  // - multiple shaders
-  // - single style
-  // - multiple styles
-  // - multiple rules
-  const asArray = Array.isArray(style) ? style : [style];
-
-  // if array of rules: break rules into separate styles, compute "else" filters
-  if ('style' in asArray[0]) {
-    /** @type {Array<StyleAsRule>} */
-    const styles = [];
-    const rules = /** @type {Array<import('../../style/flat.js').Rule>} */ (
-      asArray
-    );
-    const previousFilters = [];
-    for (const rule of rules) {
-      const ruleStyles = Array.isArray(rule.style) ? rule.style : [rule.style];
-      /** @type {import("../../expr/expression.js").EncodedExpression} */
-      let currentFilter = rule.filter;
-      if (rule.else && previousFilters.length) {
-        currentFilter = [
-          'all',
-          ...previousFilters.map((filter) => ['!', filter]),
-        ];
-        if (rule.filter) {
-          currentFilter.push(rule.filter);
-        }
-        if (currentFilter.length < 3) {
-          currentFilter = currentFilter[1];
-        }
-      }
-      if (rule.filter) {
-        previousFilters.push(rule.filter);
-      }
-      /** @type {Array<StyleAsRule>} */
-      const stylesWithFilters = ruleStyles.map((style) => ({
-        style,
-        ...(currentFilter && {filter: currentFilter}),
-      }));
-      styles.push(...stylesWithFilters);
-    }
-    return styles;
-  }
-
-  // if array of shaders: return as is
-  if ('builder' in asArray[0]) {
-    return /** @type {Array<StyleAsShaders>} */ (asArray);
-  }
-
-  return asArray.map(
-    (style) =>
-      /** @type {StyleAsRule} */ ({
-        style,
-      }),
-  );
 }
