+++ test/browser/spec/ol/render/webgl/bufferUtil.test.js
@@ -1,18 +1,13 @@
 import {
-  breakDownFlatStyle,
-  colorDecodeId,
-  colorEncodeId,
-  getBlankImageData,
   writeLineSegmentToBuffers,
   writePointFeatureToBuffers,
   writePolygonTrianglesToBuffers,
-} from '../../../../../../src/ol/render/webgl/utils.js';
+} from '../../../../../../src/ol/render/webgl/bufferUtil.js';
 import {
   compose as composeTransform,
   create as createTransform,
   makeInverse as makeInverseTransform,
 } from '../../../../../../src/ol/transform.js';
-import {ShaderBuilder} from '../../../../../../src/ol/webgl/ShaderBuilder.js';
 
 describe('webgl render utils', function () {
   describe('writePointFeatureToBuffers', function () {
@@ -489,193 +484,6 @@ describe('webgl render utils', function () {
       it('correctly returns the new index', function () {
         expect(newIndex).to.eql(29);
       });
-    });
-  });
-
-  describe('getBlankImageData', function () {
-    it('creates a 1x1 white texture', function () {
-      const texture = getBlankImageData();
-      expect(texture.height).to.eql(1);
-      expect(texture.width).to.eql(1);
-      expect(texture.data[0]).to.eql(255);
-      expect(texture.data[1]).to.eql(255);
-      expect(texture.data[2]).to.eql(255);
-      expect(texture.data[3]).to.eql(255);
-    });
-  });
-
-  describe('colorEncodeId and colorDecodeId', function () {
-    it('correctly encodes and decodes ids', function () {
-      expect(colorDecodeId(colorEncodeId(0))).to.eql(0);
-      expect(colorDecodeId(colorEncodeId(1))).to.eql(1);
-      expect(colorDecodeId(colorEncodeId(123))).to.eql(123);
-      expect(colorDecodeId(colorEncodeId(12345))).to.eql(12345);
-      expect(colorDecodeId(colorEncodeId(123456))).to.eql(123456);
-      expect(colorDecodeId(colorEncodeId(91612))).to.eql(91612);
-      expect(colorDecodeId(colorEncodeId(1234567890))).to.eql(1234567890);
-    });
-
-    it('correctly reuses array', function () {
-      const arr = [];
-      expect(colorEncodeId(123, arr)).to.be(arr);
-    });
-
-    it('is compatible with Uint8Array storage', function () {
-      const encoded = colorEncodeId(91612);
-      const typed = Uint8Array.of(
-        encoded[0] * 255,
-        encoded[1] * 255,
-        encoded[2] * 255,
-        encoded[3] * 255,
-      );
-      const arr = [
-        typed[0] / 255,
-        typed[1] / 255,
-        typed[2] / 255,
-        typed[3] / 255,
-      ];
-      const decoded = colorDecodeId(arr);
-      expect(decoded).to.eql(91612);
-    });
-  });
-
-  describe('breakDownFlatStyle', function () {
-    it('breaks down a single flat style', function () {
-      const style = {
-        'fill-color': 'red',
-        'fill-opacity': 0.5,
-        'stroke-color': 'blue',
-        'stroke-width': 2,
-      };
-      const result = breakDownFlatStyle(style);
-      expect(result).to.eql([
-        {
-          style,
-        },
-      ]);
-    });
-    it('breaks down an array of flat styles', function () {
-      const styles = [
-        {
-          'fill-color': 'red',
-          'fill-opacity': 0.5,
-        },
-        {
-          'stroke-color': 'blue',
-          'stroke-width': 2,
-        },
-      ];
-      const result = breakDownFlatStyle(styles);
-      expect(result).to.eql([
-        {
-          style: styles[0],
-        },
-        {
-          style: styles[1],
-        },
-      ]);
-    });
-    it('breaks down an array of rules, generating appropriate filters for "else" rules', function () {
-      const rules = [
-        {
-          filter: ['>', ['get', 'size'], '10'],
-          style: [
-            {'fill-color': 'red', 'fill-opacity': 0.5},
-            {'fill-color': 'green', 'fill-opacity': 0.5},
-          ],
-        },
-        {
-          else: true,
-          style: {'circle-radius': 5, 'circle-fill-color': 'red'},
-        },
-        {
-          else: true,
-          filter: ['==', ['get', 'type'], 'road'],
-          style: {'stroke-color': 'blue', 'stroke-width': 2},
-        },
-        {
-          else: true,
-          style: [
-            {'stroke-color': 'green', 'stroke-width': 2},
-            {'stroke-color': 'white', 'stroke-width': 1},
-          ],
-        },
-        {
-          style: {'stroke-color': 'yellow', 'stroke-width': 2},
-        },
-        {
-          filter: ['==', ['get', 'type'], 'street'],
-          style: {'stroke-color': 'black', 'stroke-width': 2},
-        },
-      ];
-      const result = breakDownFlatStyle(rules);
-      expect(result).to.eql([
-        {
-          filter: ['>', ['get', 'size'], '10'],
-          style: {'fill-color': 'red', 'fill-opacity': 0.5},
-        },
-        {
-          filter: ['>', ['get', 'size'], '10'],
-          style: {'fill-color': 'green', 'fill-opacity': 0.5},
-        },
-        {
-          filter: ['!', ['>', ['get', 'size'], '10']],
-          style: {'circle-radius': 5, 'circle-fill-color': 'red'},
-        },
-        {
-          filter: [
-            'all',
-            ['!', ['>', ['get', 'size'], '10']],
-            ['==', ['get', 'type'], 'road'],
-          ],
-          style: {'stroke-color': 'blue', 'stroke-width': 2},
-        },
-        {
-          'filter': [
-            'all',
-            ['!', ['>', ['get', 'size'], '10']],
-            ['!', ['==', ['get', 'type'], 'road']],
-          ],
-          style: {'stroke-color': 'green', 'stroke-width': 2},
-        },
-        {
-          'filter': [
-            'all',
-            ['!', ['>', ['get', 'size'], '10']],
-            ['!', ['==', ['get', 'type'], 'road']],
-          ],
-          style: {'stroke-color': 'white', 'stroke-width': 1},
-        },
-        {
-          style: {'stroke-color': 'yellow', 'stroke-width': 2},
-        },
-        {
-          filter: ['==', ['get', 'type'], 'street'],
-          style: {'stroke-color': 'black', 'stroke-width': 2},
-        },
-      ]);
-    });
-    it('returns an array of shaders as is', function () {
-      const shaders = [
-        {
-          builder: new ShaderBuilder(),
-          attributes: [],
-        },
-        {
-          builder: new ShaderBuilder(),
-          attributes: [],
-        },
-      ];
-      const result = breakDownFlatStyle(shaders);
-      expect(result).to.eql(shaders);
-    });
-    it('returns a single shader as array', function () {
-      const shader = {
-        builder: new ShaderBuilder(),
-        attributes: [],
-      };
-      const result = breakDownFlatStyle(shader);
-      expect(result).to.eql([shader]);
     });
   });
 });
